refactor: code organisation and naming around http routes
