+++ modules/search-api/src/main/scala/io/renku/search/api/Routes.scala
@@ -16,33 +16,45 @@
  * limitations under the License.
  */
 
-package io.renku.search.provision
+package io.renku.search.api
 
 import cats.effect.{Async, Resource}
 import cats.syntax.all.*
 import fs2.io.net.Network
+import io.renku.search.api.routes.*
 import io.renku.search.http.metrics.MetricsRoutes
 import io.renku.search.http.routes.OperationRoutes
 import io.renku.search.metrics.CollectorRegistryBuilder
+import io.renku.solr.client.SolrConfig
+import org.http4s.HttpRoutes
 import org.http4s.dsl.Http4sDsl
 import org.http4s.server.Router
-import org.http4s.{HttpApp, HttpRoutes, Response}
 
-object HttpApplication:
+private object Routes:
 
-  def apply[F[_]: Async: Network](
-      registryBuilder: CollectorRegistryBuilder[F]
-  ): Resource[F, HttpApp[F]] =
-    MetricsRoutes[F](registryBuilder).makeRoutes
-      .map(new HttpApplication[F](_).router)
+  def apply[F[_]: Async: Network](solrConfig: SolrConfig): Resource[F, HttpRoutes[F]] =
+    for
+      search <- SearchApi[F](solrConfig).map(api => SearchRoutes[F](api))
+      metricsRoutes <- MetricsRoutes[F](
+        CollectorRegistryBuilder[F].withStandardJVMMetrics
+      ).makeRoutes(search.routes)
+    yield new Routes[F](search, metricsRoutes).routes
 
-final class HttpApplication[F[_]: Async](metricsRoutes: HttpRoutes[F])
-    extends Http4sDsl[F]:
+final private class Routes[F[_]: Async](
+    searchRoutes: SearchRoutes[F],
+    metricsRoutes: HttpRoutes[F]
+) extends Http4sDsl[F]:
 
-  private lazy val operationRoutes =
+  private val prefix = "/search"
+
+  private val openapi =
+    OpenApiRoute[F](s"/api$prefix", "Renku Search API", searchRoutes.endpoints)
+
+  private lazy val searchAndOperationRoutes =
     Router[F](
+      prefix -> (openapi.routes <+> searchRoutes.routes),
       "/" -> OperationRoutes[F]
     )
 
-  lazy val router: HttpApp[F] =
-    (operationRoutes <+> metricsRoutes).orNotFound
+  lazy val routes: HttpRoutes[F] =
+    searchAndOperationRoutes <+> metricsRoutes
